Moved Request objects to `data_interface` module.
Added in comment to not delete blank __init__.py in `Tests`.
Improved handling of meld cards' backs.
