+++ Tests/TestGameMetadataCard.py
@@ -4,7 +4,7 @@ from typing import Union
 from Tests.settings import TEST_MASS_DATA_PULL
 
 from game_metadata.utils.consts import CardLayouts, CARD_SIDE
-from game_metadata.RequestScryfall import RequestScryfall
+from data_interface.RequestScryfall import RequestScryfall
 from game_metadata.game_objects.Card import Card, CardFace, CardManager
