+++ game_metadata/__init__.py
@@ -1,6 +1,5 @@
 from game_metadata.utils.consts import SUPERTYPES, TYPES, ALL_SUBTYPES, SUBTYPE_DICT, RARITY_ALIASES
 from game_metadata.utils import SETS, FORMATS
 
-from game_metadata.RequestScryfall import RequestScryfall
 from game_metadata.game_objects.Card import Card, CardManager
 from game_metadata.GameMetadata import SetMetadata, FormatMetadata
