+++ game_metadata/game_objects/Draft.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from typing import Optional
 
 from utilities.auto_logging import logging
-from game_metadata.Request17Lands import Request17Lands
+from data_interface.Request17Lands import Request17Lands
 from game_metadata.game_objects.Card import Card
 import game_metadata.game_objects.Deck as Deck
