+++ utilities/__init__.py
@@ -1,4 +1,3 @@
 from utilities.utils.settings import TRIES, FAIL_DELAY, SUCCESS_DELAY
 from utilities.utils.funcs import save_json_file, load_json_file, reformat_json_file
 from utilities import auto_logging
-from utilities.Requester import Requester
